use usize type for kmeans max_colors

+++ lib/colordom/native.rb
@@ -9,7 +9,7 @@ module Colordom
 
     attach_function :to_histogram, :to_histogram, [:string], Result
     attach_function :to_mediancut, :to_mediancut, [:string, :uint8], Result
-    attach_function :to_kmeans, :to_kmeans, [:string, :uint8], Result
+    attach_function :to_kmeans, :to_kmeans, [:string, :uint], Result
 
     attach_function :free, :free_result, [Result], :void
   end
